+++ social-net/src/Components/InputForm/module.inputForm.css
@@ -2,12 +2,13 @@
   margin: 15px 0 15px 15px;
   position: absolute;
   top: 93%;
+  display: flex;
 }
 
 .form__name {
-  margin-right: 10px;
+  margin-right: 20px;
 }
 
 .form__text {
-  margin-right: 10px;
+  margin-right: 20px;
 }
