+++ social-net/src/Components/Messeges/module.messeges.css
@@ -1,7 +1,7 @@
 .messeges {
   position: absolute;
   width: 95vw;
-  height: 90vh;
+  height: 88vh;
   overflow: auto;
 }
